test(store): port Connect tests to ava
